+++ envs/dev/views/blocks/PhpblocktestBlock.php
@@ -1,0 +1,4 @@
+#<?= $this->context->getEnvOption('index'); ?>: 
+IsPreviousEqual: <? var_dump($this->context->getEnvOption('isPrevEqual')); ?>
+IsNextEqual: <? var_dump($this->context->getEnvOption('isNextEqual')); ?>
+<hr />