+++ modules/cmsadmin/src/base/PhpBlockInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace cmsadmin\base;
+
+interface PhpBlockInterface
+{
+    public function frontend();
+    
+    public function admin();
+}