+++ modules/cmsadmin/src/base/TwigBlockInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace cmsadmin\base;
+
+interface TwigBlockInterface
+{
+    public function twigFrontend();
+    
+    public function twigAdmin();
+    
+    public function render();
+}